+++ packages/expo-brownfield/cli/src/commands/build-ios.ts
@@ -5,6 +5,7 @@ import {
   printIosConfig,
   resolveBuildConfigIos,
   validatePrebuild,
+  validatePrecompiledModules,
   shipSwiftPackage,
   shipFrameworks,
 } from '../utils';
@@ -12,6 +13,9 @@ import {
 const buildIos = async (command: Command) => {
   await validatePrebuild('ios');
   const config = resolveBuildConfigIos(command.opts());
+  if (config.usePrebuilds) {
+    validatePrecompiledModules();
+  }
   printIosConfig(config);
 
   await buildFramework(config);